refactor(WelcomeScreen): use a local function in ProjectErrorMoreInfo
We don't want to write "TextBlockDisplayStyle.WrappedText" for every
line, so we use a local function to do it for us.

Also, remove gui.allocator = RectAllocator.LeftAlign; because we use
the default of TextBlockDisplayStyle -
RectAlignment Alignment = RectAlignment.MiddleLeft.
It's not an identical action, but ImGui#AllocateRect returns the
rectangle without any alignment if we use gui.allocator, which
likely is not what we want in this situation.

On top of that, I don't endorse using a field from a god class.
We want to refactor such occasions to something more understandable
to a new reader.